IMU_processing/FFT/requirements.txt
core==1.0.1
keras==2.10.0
matplotlib==3.6.1
numpy==1.23.4
pandas==1.4.4
Pillow==9.3.0
pydot==1.4.2
scikit_learn==1.1.3
scipy==1.9.3
tensorflow==2.10.0
tensorflow_gpu==2.10.0
torch==1.13.0
